+++ R/yamlR.R
@@ -231,13 +231,15 @@ yml_write <- function(config, path = NULL) {
 #' yml_write(conf,path=tmp)
 #' @export
 construct_yml_config <- function(code = NULL, data = NULL, render_root = NULL) {
-  code <- basename(code)
+  if (!is.null(code)) {
+    code <- basename(code)
+  }
   files <- vector(length = length(code), mode = "list")
   names(files) <- code
   for (i in code) {
-    # files[[i]]$name <- i
     files[[i]]$enabled <- TRUE
   }
+  
   # create render root at a temporary directory.
   # this will be stored in the yaml. What if we restart?
   # see processData - it gets validated and created if not existing.
@@ -260,7 +262,7 @@ construct_yml_config <- function(code = NULL, data = NULL, render_root = NULL) {
       silent = TRUE
       )
     if (inherits(render_root, "try-error")) {
-      flog.fatal(paste0(
+      .multilog_fatal(paste0(
         dirname(render_root),
         " doesn't exist!"
       ))
@@ -279,7 +281,7 @@ construct_yml_config <- function(code = NULL, data = NULL, render_root = NULL) {
   } else if (length(x$configuration$render_root) != 0) {
     return(x$configuration$render_root)
   } else {
-    flog.fatal("render_root is not set in yaml")
+    .multilog_fatal("render_root is not set in yaml")
     stop("error", call. = FALSE)
   }
 }